* simplify  ;-)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@19529 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/net/Pack.cxx
@@ -298,7 +298,7 @@ void* nboUnpackI64(void* b, int64_t& v)
   if (ErrorChecking) {
     if (Length < sizeof(int64_t)) {
       Error = true;
-      v = (int64_t) 0.0f;
+      v = 0;
       return b;
     } else {
       Length -= sizeof(int64_t);
@@ -316,7 +316,7 @@ void* nboUnpackU64(void* b, uint64_t& v)
   if (ErrorChecking) {
     if (Length < sizeof(uint64_t)) {
       Error = true;
-      v = (uint64_t) 0.0f;
+      v = 0;
       return b;
     } else {
       Length -= sizeof(uint64_t);
